# docs/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "独自コインAPI"
    }
  ],
  "paths": {
    "/api/user": {
      "post": {
        "tags": [
          "独自コインAPI"
        ],
        "summary": "ユーザ登録API",
        "description": "独自コインを利用するためのユーザ登録をおこなう",
        "operationId": "registerUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/jp.gr.java_conf.gtask.domain.datetime.RequestDateTime"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.registeruser.response.RegisterUserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/{userId}/transfer": {
      "post": {
        "tags": [
          "独自コインAPI"
        ],
        "summary": "送金API",
        "description": "他のユーザへ独自コインの送金をおこなう",
        "operationId": "transfer",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "requestDateTime",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/jp.gr.java_conf.gtask.domain.datetime.RequestDateTime"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.transfer.request.TransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.transfer.response.TransferResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/{userId}/payment": {
      "post": {
        "tags": [
          "独自コインAPI"
        ],
        "summary": "支払いAPI",
        "description": "独自コインを使用して支払いをおこなう",
        "operationId": "payment",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "requestDateTime",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/jp.gr.java_conf.gtask.domain.datetime.RequestDateTime"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.payment.request.PaymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/{userId}/history": {
      "post": {
        "tags": [
          "独自コインAPI"
        ],
        "summary": "履歴API",
        "description": "出入金の履歴を取得する",
        "operationId": "gethistory",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.gethistory.response.GetHistoryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/{userId}/balance": {
      "get": {
        "tags": [
          "独自コインAPI"
        ],
        "summary": "残高取得API",
        "description": "ユーザーの残高を返却します",
        "operationId": "getBalance",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.getbalance.response.GetBalanceResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "独自コインAPI"
        ],
        "summary": "入金API",
        "description": "独自コインに入金をおこなう",
        "operationId": "addBalance",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "requestDateTime",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/jp.gr.java_conf.gtask.domain.datetime.RequestDateTime"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.addbalance.request.AddBalanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.addbalance.response.AddBalanceResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "jp.gr.java_conf.gtask.domain.datetime.RequestDateTime": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.registeruser.response.ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.registeruser.response.RegisterUserResponse": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.registeruser.response.ResultResponse"
          },
          "error": {
            "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.registeruser.response.ErrorResponse"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.registeruser.response.ResultResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.transfer.request.TransferRequest": {
        "type": "object",
        "properties": {
          "recipientUserId": {
            "type": "integer",
            "format": "int64"
          },
          "balance": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.transfer.response.ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.transfer.response.ResultResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "balance": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.transfer.response.TransferResponse": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.transfer.response.ResultResponse"
          },
          "error": {
            "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.transfer.response.ErrorResponse"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.payment.request.PaymentRequest": {
        "type": "object",
        "properties": {
          "balance": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.gethistory.response.ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.gethistory.response.GetHistoryResponse": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.gethistory.response.ResultResponse"
          },
          "error": {
            "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.gethistory.response.ErrorResponse"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.gethistory.response.HistoryResponse": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string"
          },
          "balance": {
            "type": "integer",
            "format": "int64"
          },
          "updateDate": {
            "type": "string"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.gethistory.response.ResultResponse": {
        "type": "object",
        "properties": {
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.gethistory.response.HistoryResponse"
            }
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.addbalance.request.AddBalanceRequest": {
        "type": "object",
        "properties": {
          "balance": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.addbalance.response.AddBalanceResponse": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.addbalance.response.ResultResponse"
          },
          "error": {
            "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.addbalance.response.ErrorResponse"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.addbalance.response.ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.addbalance.response.ResultResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "balance": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.getbalance.response.ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.getbalance.response.GetBalanceResponse": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.getbalance.response.ResultResponse"
          },
          "error": {
            "$ref": "#/components/schemas/jp.gr.java_conf.gtask.presentation.user.getbalance.response.ErrorResponse"
          }
        }
      },
      "jp.gr.java_conf.gtask.presentation.user.getbalance.response.ResultResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "balance": {
            "type": "integer",
            "format": "int64"
          }
        }
      }
    }
  }
}